+++ poisson_solver.hpp
@@ -7,23 +7,23 @@ class PoissonSolver
 {
 private:
     const int n;
-    const std::vector<double> &source;
+    const std::vector<float> &source;
     const int iterations;
     const int threads;
     const float delta;
     bool debug = false;
-    std::vector<double> *curr;
-    std::vector<double> *next;
+    std::vector<float> *curr;
+    std::vector<float> *next;
 
 public:
     PoissonSolver(int n,
-                  const std::vector<double> &source,
+                  const std::vector<float> &source,
                   int iterations,
                   int threads,
                   float delta,
                   bool debug);
 
-    std::vector<double> *solve(void);
+    std::vector<float> *solve(void);
 };
 
 #endif
